api/requirements.txt
aniso8601==7.0.0
APScheduler==3.6.1
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Authlib==0.12.1
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-JWT-Extended==3.23.0
flask-restplus==0.13.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
lazy-object-proxy==1.4.2
loginpass==0.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
pathlib==1.0.1
peewee==3.10.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
pyunifi==2.18
PyYAML==5.1.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.6
Werkzeug==0.15.6
wrapt==1.11.2
wtf-peewee==3.0.0
yapf==0.28.0
